Database.PostgreSQL.Simple.Missing

Documentation

intervalSeconds :: Parser Pico

interval :: Parser (Int, Int, Pico)

Parse a limited postgres interval of the form [-]HHH:MM:SS.[SSSS] (no larger units than hours).

data T

Constructors

T !Int !Int64

seconds :: Parser Pico

Parse a count of seconds, with the integer part being two digits long.

twoDigits :: Parser Int

Parse a two-digit integer (e.g. day of month, hour).