Thentos

Documentation

main :: IO ()

makeMain :: (ActionState -> Maybe HttpConfig -> Maybe HttpConfig -> IO ()) -> IO ()

createConnPoolAndInitDb :: SBS -> IO (Pool Connection)

Create a connection pool and initialize the DB by creating all tables, indexes etc. if the DB is empty. Tables already existing in the DB won't be touched. The DB itself must already exist.

createDefaultUser :: Connection -> Maybe DefaultUserConfig -> IO ()

If default user is Nothing or user with 'UserId 0' exists, do nothing. Otherwise, create default user.

autocreateMissingServices :: ThentosConfig -> Action Void s ()

Autocreate any services that are listed in the config but don't exist in the DB. Dies with an error if the default "proxy" service ID is repeated in the "proxies" section.