Thentos.Action.SimpleAuth

Description

Simplified access to Action with guarded exits.

Documentation

newtype UnsafeAction e s a

Like Action, but with IO at the base.

Constructors

UnsafeAction

Fields

fromUnsafeAction :: ReaderT ActionState (EitherT (ThentosError e) (StateT s IO)) a

assertAuth :: Action e s Bool -> Action e s ()

Run boolean authorization predicate. Throw ActionErrorAnyLabel if the result is False.

hasAgent :: Agent -> Action e s Bool

hasUserId :: UserId -> Action e s Bool

hasServiceId :: ServiceId -> Action e s Bool

hasRole :: Role -> Action e s Bool

guardedUnsafeAction :: Action e s Bool -> UnsafeAction e s a -> Action e s a

Run an UnsafeAction in a safe Action with extra authorization checks (performed through assertAuth).

unsafeAction :: forall e s a. UnsafeAction e s a -> Action e s a

Run an UnsafeAction in a safe Action without extra authorization checks.