Thentos.Frontend.Pages

Description

This module provides Html pages and Forms (disgetive-functor thingies that parse filled-out forms). The pages come in three flavours:

*Pages* (starting with html tag and without holes);

*Pagelets* (functions that return *complete pages*, but contain holes that need to be filled with the function's Html argument(s));

*Snippets*: Html elements for filling holes in pagelets or other snippets.

Documentation

dashboardPagelet :: FrontendSessionData -> [Role] -> Html -> Html

The dashboard is the frame of what the user always sees when logged in. The dashboard body shows further specifics. It is the caller's responsibility to make sure that dashboard state and body correspond.

userRegisterPage :: FrontendSessionData -> View Html -> ST -> Html

userRegisterForm :: Monad m => Form Html m UserFormData

userRegisterRequestedPage :: FrontendSessionData -> Html

userLoginPage :: FrontendSessionData -> View Html -> ST -> Html

userLoginForm :: Monad m => Form Html m (UserName, UserPass)

resetPasswordRequestPage :: FrontendSessionData -> View Html -> ST -> Html

resetPasswordRequestForm :: Monad m => Form Html m UserEmail

resetPasswordRequestedPage :: FrontendSessionData -> Html

resetPasswordPage :: FrontendSessionData -> View Html -> ST -> Html

resetPasswordForm :: Monad m => Form Html m UserPass

userLogoutConfirmSnippet :: ST -> [ServiceName] -> ST -> u -> rs -> Html

userLogoutDonePage :: FrontendSessionData -> Html

userDisplaySnippet :: User -> rs -> Html

userServicesDisplaySnippet :: u -> rs -> Html

(this is just a dummy.)

emailUpdateSnippet :: FrontendSessionData -> View Html -> ST -> u -> rs -> Html

emailUpdateForm :: Monad m => Form Html m UserEmail

passwordUpdateSnippet :: FrontendSessionData -> View Html -> ST -> u -> rs -> Html

passwordUpdateForm :: Monad m => Form Html m (UserPass, UserPass)

serviceCreateSnippet :: FrontendSessionData -> View Html -> ST -> u -> rs -> Html

serviceCreateForm :: Monad m => Form Html m (ServiceName, ServiceDescription)

serviceRegisterPage :: FrontendSessionData -> View Html -> ST -> ServiceId -> Service -> User -> Html

serviceRegisterForm :: Monad m => Form Html m ()

errorPage :: String -> Html

errorPagelet :: u -> rs -> String -> Html

permissionDeniedPage :: Html

notFoundPage :: Html

confirmationMailSentPage :: FrontendSessionData -> ST -> ST -> ST -> Html

confirmationMailSentSnippet :: ST -> ST -> u -> rs -> Html