Thentos.Frontend.Types

content types

type HTM = PrettyHTML

Html content type with pretty printing. (See also: package servant-blaze.)

renderHTM :: Html -> LBS

data PrettyHTML

data TextCss

frontend actions

type FAction = Action FActionError FrontendSessionData

data FActionError

Constructors

FActionError303 SBS

FActionError404

FActionErrorNoToken

FActionErrorCreateService

FActionErrorServiceLoginNoCbUrl

FActionError500 String

session state

data FrontendSessionData

Constructors

FrontendSessionData

Fields

_fsdLogin :: Maybe FrontendSessionLoginData

_fsdServiceLoginState :: Maybe ServiceLoginState

(see ServiceLoginState)

_fsdMessages :: [FrontendMsg]

emptyFrontendSessionData :: FrontendSessionData

data FrontendSessionLoginData

Constructors

FrontendSessionLoginData

Fields

_fslToken :: ThentosSessionToken

_fslUserId :: UserId

_fslDashboardTab :: Maybe DashboardTab

data DashboardTab

Constructors

DashboardTabDetails

DashboardTabServices

DashboardTabOwnServices

DashboardTabUsers

DashboardTabLogout

data ServiceLoginState

If a user comes from a service login and is sent to the "register with a new service" page because no valid ServiceAccount exists in thentos, an extra round of indirection is required (the user is confronted with the service details and must say "yes, i want to register", "share this-and-that data", etc.). There may be further extra rounds for logging in to thentos or registering with thentos. This type is used to store the information from the service login request in FrontendSessionData until all these steps have been taken.

Constructors

ServiceLoginState

Fields

_fslServiceId :: ServiceId

_fslRR :: RelativeRef

e.g. servicelogin?... FIXME: give fslRR a better name. something that explains the meaning of where it is pointing.

data FrontendMsg

If we want to communicate information from the last request to the user in the next one, we need to stash it in the state. This is the type for that.

Constructors

FrontendMsgError ST

FrontendMsgSuccess ST

fsdServiceLoginState :: Lens' FrontendSessionData (Maybe ServiceLoginState)

fsdMessages :: Lens' FrontendSessionData [FrontendMsg]

fsdLogin :: Lens' FrontendSessionData (Maybe FrontendSessionLoginData)

fslUserId :: Lens' FrontendSessionLoginData UserId

fslToken :: Lens' FrontendSessionLoginData ThentosSessionToken

fslDashboardTab :: Lens' FrontendSessionLoginData (Maybe DashboardTab)

fslServiceId :: Lens' ServiceLoginState ServiceId

fslRR :: Lens' ServiceLoginState RelativeRef