Thentos.Types

Documentation

data User

Constructors

User

Fields

_userName :: !UserName

_userPassword :: !(HashedSecret UserPass)

_userEmail :: !UserEmail

data ServiceAccount

the data a user maintains about a service they are signed up with.

Constructors

ServiceAccount

Fields

_serviceAnonymous :: !Bool

Do not give out any information about user beyond session token validity bit. (not implemented.)

newServiceAccount :: ServiceAccount

newtype UserId

Constructors

UserId

Fields

fromUserId :: Integer

newtype UserName

Constructors

UserName

Fields

fromUserName :: ST

newtype UserPass

FIXME: ToJSON instance should go away in order to avoid accidental leakage of cleartext passwords. but for the experimentation phase this is too much of a headache. (Under no circumstances render to something like "[password hidden]". Causes a lot of confusion.)

Constructors

UserPass

Fields

fromUserPass :: ST

newtype HashedSecret a

Constructors

HashedSecret

Fields

fromHashedSecret :: EncryptedPass

newtype UserEmail

Constructors

UserEmail

Fields

userEmailAddress :: EmailAddress

parseUserEmail :: ST -> Maybe UserEmail

fromUserEmail :: UserEmail -> ST

newtype ConfirmationToken

Constructors

ConfirmationToken

Fields

fromConfirmationToken :: ST

newtype PasswordResetToken

Constructors

PasswordResetToken

Fields

fromPasswordResetToken :: ST

data UserFormData

Information required to create a new User

Constructors

UserFormData

Fields

udName :: !UserName

udPassword :: !UserPass

udEmail :: !UserEmail

data LoginFormData

Constructors

LoginFormData

Fields

ldName :: !UserName

ldPassword :: !UserPass

data Service

(Service owner is an Agent, not a User, so that services can (but do not have to) be owned by their parent services in a service hierarchy.)

Constructors

Service

Fields

_serviceKey :: !(HashedSecret ServiceKey)

_serviceOwner :: !UserId

_serviceThentosSession :: !(Maybe ThentosSessionToken)

Used by the service to authenticate in communication with thentos.

_serviceName :: !ServiceName

_serviceDescription :: !ServiceDescription

newtype ServiceId

Constructors

ServiceId

Fields

fromServiceId :: ST

newtype ServiceKey

Constructors

ServiceKey

Fields

fromServiceKey :: ST

newtype ServiceName

Constructors

ServiceName

Fields

fromServiceName :: ST

newtype ServiceDescription

Constructors

ServiceDescription

Fields

fromServiceDescription :: ST

newtype Group

Service-side authoriziation classes. (For thentos-internal authorization classes, see Role.)

Groups are opaque strings that services can use to manage authorizations for their users in thentos. One reason why thentos offers this (rather than leaving the groups-to-users mapping to the internals of the service) is that this puts us in a position to do anonymized authentication: we can assert a request is issued by a user member in a certain group, but not leak the name of the user.

Constructors

Group

Fields

fromGroup :: ST

newtype PersonaId

Constructors

PersonaId

Fields

fromPersonaId :: Integer

newtype PersonaName

Constructors

PersonaName

Fields

fromPersonaName :: ST

data Persona

Note on the external url field:* Since personas are exposed to the service, it sometimes makes sense for a service to maintain its own data item for each persona in thentos. The persona's external url can be used to point to that data item's rest url.

Constructors

Persona

Fields

_personaId :: PersonaId

_personaName :: PersonaName

_personaUid :: UserId

_personaExternalUrl :: Maybe Uri

newtype ContextId

Constructors

ContextId

Fields

fromContextId :: Integer

newtype ContextName

Constructors

ContextName

Fields

fromContextName :: ST

newtype ContextDescription

Constructors

ContextDescription

Fields

fromContextDescription :: ST

data Context

Constructors

Context

Fields

_contextId :: ContextId

_contextService :: ServiceId

_contextName :: ContextName

_contextDescription :: ContextDescription

_contextUrl :: Maybe ProxyUri

newtype ThentosSessionToken

Constructors

ThentosSessionToken

Fields

fromThentosSessionToken :: ST

data ThentosSession

Constructors

ThentosSession

Fields

_thSessAgent :: !Agent

_thSessStart :: !Timestamp

_thSessEnd :: !Timestamp

_thSessExpirePeriod :: !Timeout

newtype ServiceSessionToken

Constructors

ServiceSessionToken

Fields

fromServiceSessionToken :: ST

data ServiceSession

Constructors

ServiceSession

Fields

_srvSessService :: !ServiceId

_srvSessStart :: !Timestamp

_srvSessEnd :: !Timestamp

_srvSessExpirePeriod :: !Timeout

_srvSessThentosSession :: !ThentosSessionToken

_srvSessMetadata :: !ServiceSessionMetadata

data ServiceSessionMetadata

Constructors

ServiceSessionMetadata

Fields

_srvSessMdUser :: !UserName

data ByUserOrServiceId

Constructors

ByUser (UserId, UserPass)

ByService (ServiceId, ServiceKey)

newtype Timestamp

Constructors

Timestamp

Fields

fromTimestamp :: UTCTime

newtype Timeout

Constructors

Timeoutms

Fields

toMilliseconds :: Int

toSeconds :: (Fractional a, Real a) => Timeout -> a

fromMilliseconds :: Int -> Timeout

fromSeconds :: Int -> Timeout

fromMinutes :: Int -> Timeout

fromHours :: Int -> Timeout

fromDays :: Int -> Timeout

timestampToString :: Timestamp -> String

timestampFromString :: Monad m => String -> m Timestamp

timeoutToString :: Timeout -> String

timeoutFromString :: Monad m => String -> m Timeout

secondsToString :: (Show a, Fractional a, Real a) => a -> String

secondsFromString :: (Read a, Fractional a, Real a, Monad m) => String -> m a

data Agent

Some thing or body that deals with (and can authenticate itself before) thentos. Examples: User or Service. (We could have called this Principal, but that name is in use by LIO already.)

Constructors

UserA !UserId

ServiceA !ServiceId

data Role

Thentos-internal authorization classes. (See Group for service-side authorization classes.)

Constructors

RoleAdmin
Can do anything. (There may be no difference in behaviour from allowEverything resp. thentosPublic, but if we ever want to restrict privileges, it's easier if it is a Role.)

RoleUser
Can sign up with services

RoleUserAdmin
Can create (and manage her own) users

RoleServiceAdmin
Can create (and manage her own) services

RoleGroupAdmin
Can add personas and groups to groups and remove them

newtype RelRef

Constructors

RelRef

Fields

fromRelRef :: RelativeRef

newtype Uri

Wrapper around URI with additional instance definitions.

Constructors

Uri

Fields

fromUri :: URI

parseUri :: SBS -> Either URIParseError Uri

renderUri :: Uri -> SBS

data ProxyUri

Constructors

ProxyUri

Fields

proxyHost :: SBS

proxyPort :: Int

proxyPath :: SBS

renderProxyUri :: ProxyUri -> ST

parseProxyUri :: forall m. MonadError String m => ST -> m ProxyUri

(<//>) :: (ConvertibleStrings s ST, ConvertibleStrings ST s) => s -> s -> s

Path concatenation for avoiding double slashes in paths. One optional / trailing left side / leading right side is removed, and one / is inserted.

stripLeadingSlash :: ST -> ST

Strip an optional slash from the start of a text. If the text doesn't start with a slash, it is returned unchanged.

stripTrailingSlash :: ST -> ST

Strip an optional slash from the end of a text. If the text doesn't end in a slash, it is returned unchanged.

data Random20

20 bytes of randomness. For comparison: an UUID has 16 bytes, so that should be enough for all practical purposes.

mkRandom20 :: SBS -> Maybe Random20

Construct a Random20 from a bytestring. Returns Just a Random20 wrapping the input if its length is 20, Nothing otherwise.

fromRandom20 :: Random20 -> SBS

Extract the wrapped 20 bytes from a Random20.

newtype ImageData

Constructors

ImageData

Fields

fromImageData :: SBS

newtype CaptchaId

Constructors

CaptchaId

Fields

fromCaptchaId :: ST

data ThentosError e

Constructors

NoSuchUser

NoSuchPendingUserConfirmation

MalformedConfirmationToken ST

ConfirmationTokenAlreadyExists

NoSuchService

NoSuchThentosSession

NoSuchServiceSession

NoSuchPersona

NoSuchContext

MultiplePersonasPerContext

GroupMembershipLoop Group Group

OperationNotPossibleInServiceSession

ServiceAlreadyExists

NotRegisteredWithService

UserEmailAlreadyExists

UserNameAlreadyExists

UserIdAlreadyExists

PersonaNameAlreadyExists

ContextNameAlreadyExists

CaptchaIdAlreadyExists

NoSuchCaptchaId

BadCredentials

BadAuthenticationHeaders

ProxyNotAvailable

MissingServiceHeader

ProxyNotConfiguredForService ServiceId

NoSuchToken

NeedUserA ThentosSessionToken ServiceId

MalformedUserPath ST

OtherError e

personaId :: Lens' Persona PersonaId

personaName :: Lens' Persona PersonaName

personaUid :: Lens' Persona UserId

personaExternalUrl :: Lens' Persona (Maybe Uri)

contextDescription :: Lens' Context ContextDescription

contextId :: Lens' Context ContextId

contextName :: Lens' Context ContextName

contextService :: Lens' Context ServiceId

contextUrl :: Lens' Context (Maybe ProxyUri)

serviceDescription :: Lens' Service ServiceDescription

serviceKey :: Lens' Service (HashedSecret ServiceKey)

serviceName :: Lens' Service ServiceName

serviceOwner :: Lens' Service UserId

serviceThentosSession :: Lens' Service (Maybe ThentosSessionToken)

serviceAnonymous :: Iso' ServiceAccount Bool

srvSessEnd :: Lens' ServiceSession Timestamp

srvSessExpirePeriod :: Lens' ServiceSession Timeout

srvSessMetadata :: Lens' ServiceSession ServiceSessionMetadata

srvSessService :: Lens' ServiceSession ServiceId

srvSessStart :: Lens' ServiceSession Timestamp

srvSessThentosSession :: Lens' ServiceSession ThentosSessionToken

thSessAgent :: Lens' ThentosSession Agent

thSessEnd :: Lens' ThentosSession Timestamp

thSessExpirePeriod :: Lens' ThentosSession Timeout

thSessStart :: Lens' ThentosSession Timestamp

userEmail :: Lens' User UserEmail

userName :: Lens' User UserName

userPassword :: Lens' User (HashedSecret UserPass)