Thentos.Util

Documentation

hashUserPass :: (Functor m, MonadIO m) => UserPass -> m (HashedSecret UserPass)

hashServiceKey :: (Functor m, MonadIO m) => ServiceKey -> m (HashedSecret ServiceKey)

secretMatches :: ST -> HashedSecret a -> Bool

verifyPass :: UserPass -> User -> Bool

verifyKey :: ServiceKey -> Service -> Bool

makeUserFromFormData :: (Functor m, MonadIO m) => UserFormData -> m User

mailEncode :: ConvertibleStrings s SBS => s -> ST

Encode a bytestring in such a way that it can be used as local part in an email address. This is done by percent-encoding the input in such a way that it could be used in a query string and additionally replacing every "." by "+", since the local address part cannot contain multiple subsequent dots or start or end with a dot.

cshow :: (Show a, ConvertibleStrings String b) => a -> b

Convertible show.

readsPrecEnumBoundedShow :: (Enum a, Bounded a, Show a) => Int -> String -> [(a, String)]

Generic readsPrec for enumerable types.

fmapLM :: (Monad m, Functor m) => (a -> m b) -> Either a r -> m (Either b r)

Like fmapL from Data.EitherR, but with the update of the left value constructed in an impure action.

fmapLTM :: (Monad m, Functor m) => (a -> m b) -> EitherT a m r -> EitherT b m r

Like fmapLT from Data.EitherR, but with the update of the left value constructed in an impure action.